Implementation with interfaces
TODO : now that Cards are no longer unique (ie an object is created each time), we need to store them somewhere, to keep track of which is which. We will then save the hashCode of the card in the ItemStack, so accessing the ItemStack will give us (after checking the stored cards) the actual unique card object related.